refactor rankedplay related code
