+++ packages/tosu/src/memory/lazer.ts
@@ -66,10 +66,6 @@ import {
     ModsAcronyms,
     ModsCategories
 } from '@/utils/osuMods.types';
-import {
-    readNullableInt,
-    readSharpDictionaryIntToRef
-} from '@/utils/tprocessExtensions';
 
 type LazerPatternData = {
     scalingContainerTargetDrawSize: number;
@@ -3642,7 +3638,7 @@ export class LazerMemory extends AbstractMemory<LazerPatternData> {
         return { chat: chatItems, spectatingClients };
     }
 
-    rankedPlay(): IRankedPlay | 'not-ready' {
+    rankedPlay(): IRankedPlay {
         const multiplayerClient = this.multiplayerClient();
 
         if (!multiplayerClient) {
@@ -3723,10 +3719,7 @@ export class LazerMemory extends AbstractMemory<LazerPatternData> {
 
         // i believe its proved this state is rankedplay state
 
-        const users = readSharpDictionaryIntToRef(
-            this.process,
-            usersDictionary
-        );
+        const users = this.process.readSharpDictionaryIntToRef(usersDictionary);
 
         return {
             stage,
@@ -3762,15 +3755,13 @@ export class LazerMemory extends AbstractMemory<LazerPatternData> {
                     )
                 }
             })),
-            activeUserId: readNullableInt(
-                this.process,
+            activeUserId: this.process.readNullableInt(
                 state +
                     this.offsets[
                         'osu.Game.Online.Multiplayer.MatchTypes.RankedPlay.RankedPlayRoomState'
                     ]['<ActiveUserId>k__BackingField']
             ),
-            winningUserId: readNullableInt(
-                this.process,
+            winningUserId: this.process.readNullableInt(
                 state +
                     this.offsets[
                         'osu.Game.Online.Multiplayer.MatchTypes.RankedPlay.RankedPlayRoomState'
